auth_service/pyproject.toml
[tool.poetry]
name = "auth_service"
version = "0.1.0"
description = "Authentication microservice"
authors = ["LonerLykos <loner.lykos.project@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
shared = {path = "../shared", develop = false}
sqlalchemy = "^2.0.41"
asyncmy = "^0.2.10"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.16.1"
uvicorn = "^0.34.3"
email-validator = "^2.2.0"
python-multipart = "^0.0.20"
bcrypt = "^4.3.0"
typer = "^0.16.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"
pytest-mock = "^3.14.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
